+++ metricflow/sql_clients/duckdb.py
@@ -12,12 +12,12 @@ from typing_extensions import override
 
 from metricflow.dataflow.sql_table import SqlTable
 from metricflow.protocols.sql_client import SqlEngine
-from metricflow.protocols.sql_request import SqlJsonTag, SqlRequestTagSet
 from metricflow.sql.render.duckdb_renderer import DuckDbSqlQueryPlanRenderer
 from metricflow.sql.render.sql_plan_renderer import SqlQueryPlanRenderer
 from metricflow.sql.sql_bind_parameters import SqlBindParameters
 from metricflow.sql_clients.common_client import SqlDialect
 from metricflow.sql_clients.sqlalchemy_dialect import SqlAlchemySqlClient
+from metricflow.sql_request.sql_request_attributes import SqlJsonTag, SqlRequestTagSet
 
 logger = logging.getLogger(__name__)
