+++ metricflow/sql_clients/sqlalchemy_dialect.py
@@ -11,9 +11,9 @@ import sqlalchemy
 from sqlalchemy import inspect
 
 from metricflow.dataflow.sql_table import SqlTable
-from metricflow.protocols.sql_request import SqlJsonTag, SqlRequestTagSet
 from metricflow.sql.sql_bind_parameters import SqlBindParameters
 from metricflow.sql_clients.base_sql_client_implementation import BaseSqlClientImplementation
+from metricflow.sql_request.sql_request_attributes import SqlJsonTag, SqlRequestTagSet
 
 logger = logging.getLogger(__name__)
